+++ adapt/gcloud/index.tsx
@@ -1,0 +1,1 @@
+export { CloudRun } from "./CloudRun";
